+++ projects/ui-lib/content/mosaic-parallax/mosaic-parallax.interface.ts
@@ -1,0 +1,7 @@
+import { UiLibImageI } from "@lluc_llull/ui-lib/interfaces";
+
+export interface MosaicParallaxI{
+  title?: string;
+  text?: string;
+  images?: UiLibImageI[];
+}